feat(buildkit): simplify build process by integrating build logic into index.js
