+++ .changeset/wicked-walls-camp.md
@@ -1,0 +1,5 @@
+---
+"@withstudiocms/buildkit": patch
+---
+
+Simplify buildkit
